+++ Project/BackEnd/Models/Rental.cs
@@ -14,7 +14,7 @@ public class Rental_Car
     public string Model { get; set; }
 
     [BsonElement("price_per_day")]
-    public int Price_Per_Day { get; set; }
+    public double Price_Per_Day { get; set; }
 }
 
 public class Customer
@@ -32,12 +32,15 @@ public class Customer
 public class Rental_Details
 {
     [BsonElement("start_date")]
+    // [BsonDateTimeOptions(DateOnly = true)]
     public DateTime Start_Date { get; set; }
 
     [BsonElement("expected_end_date")]
+    // [BsonDateTimeOptions(DateOnly = true)]
     public DateTime Expected_End_Date { get; set; }
 
     [BsonElement("end_date")]
+    // [BsonDateTimeOptions(DateOnly = true)]
     public DateTime? End_Date { get; set; }
 
     [BsonElement("rental_status")]
@@ -47,37 +50,37 @@ public class Rental_Details
     public string Insurance_Type { get; set; }
 
     [BsonElement("extra_insurance_amount")]
-    public int Extra_Insurance_Amount { get; set; }
+    public double Extra_Insurance_Amount { get; set; }
 
     [BsonElement("days")]
     public int Days { get; set; }
 
     [BsonElement("extra_days_amount")]
-    public int Extra_Days_Amount { get; set; }
+    public double Extra_Days_Amount { get; set; }
 
     [BsonElement("mileage")]
     public int Mileage { get; set; }
 
     [BsonElement("extra_mileage_amount")]
-    public int Extra_Mileage_Amount { get; set; }
+    public double Extra_Mileage_Amount { get; set; }
 
     [BsonElement("extra_fuel")]
     public int? Extra_Fuel { get; set; }
 
     [BsonElement("extra_fuel_amount")]
-    public int Extra_Fuel_Amount { get; set; }
+    public double Extra_Fuel_Amount { get; set; }
 
     [BsonElement("price")]
-    public int Price { get; set; }
+    public double Price { get; set; }
 
     [BsonElement("discount")]
     public double Discount { get; set; }
 
     [BsonElement("extra_amount")]
-    public int Extra_Amount { get; set; }
+    public double Extra_Amount { get; set; }
 
     [BsonElement("final_amount")]
-    public int Final_Amount { get; set; }
+    public double Final_Amount { get; set; }
 }
 
 public class Rental
